refactor(meta-service): remove redundant ActionHandler; move logic into MetaServiceImpl
